Relocate rerun-trigger family to photon_feedback_derive (Phase 8 batch 2, #688)
Issue #688 (parent #680, Phase 8) batch 2: migrate the rerun-trigger
keyword detector family from `turn.rs` to `photon_feedback_derive.rs`.

Migrated as free fns in `photon_feedback_derive.rs`:

* `pub(crate) fn is_rerun_trigger(msg) -> bool`
* `fn normalize_rerun_trigger_input(msg) -> String` (module-private)
* `fn contains_rerun_with_word_boundary(s) -> bool` (module-private)
* `pub(crate) fn build_rerun_prompt_hint_if_eligible(
    user_msg, &SessionSnapshot) -> Option<String>`
* `pub(super) fn is_runnable_rerun_hint(cmd) -> bool` (was
  module-private; promoted to `pub(super)` so the in-crate
  `rerun_hint_eligibility_tests` test mod in `turn.rs` can keep
  testing it via `super::super::photon_feedback_derive::*`)

`photon_feedback_derive.rs` imports gain `super::completion_evidence`,
`crate::session::feedback::MAX_VERIFIER_COMMAND_BYTES`,
`crate::session::store::SessionSnapshot`,
`crate::tools::bash::{BashCommandClass, classify_command}`.

`turn.rs` adjustments:

* Removed the 5 fn definitions (originally at lines 552-686 + 728-752).
* Added `is_rerun_trigger` / `build_rerun_prompt_hint_if_eligible`
  to the existing `use super::photon_feedback_derive::{...}` block
  (the `is_rerun_trigger` import was dropped in the same edit since
  the only production caller is `build_rerun_prompt_hint_if_eligible`
  which now lives in `photon_feedback_derive.rs` itself; only the
  test mod needed the explicit `super::super::photon_feedback_derive::*`
  path).
* Test mods (`is_rerun_trigger_tests`, `rerun_hint_eligibility_tests`)
  rewrote their imports to `super::super::photon_feedback_derive::*`.

LOC delta:

* `turn.rs`: 29,808 → 29,629 (-179 LOC)
* `photon_feedback_derive.rs`: 141 → 320 (+179 LOC)
* Cumulative `turn.rs`: 39,489 → 29,629 (-9,860 LOC, -25.0%)

`pub(crate)` for the public-facing rerun trigger fns (matching the
pre-extraction visibility), `pub(super)` for `is_runnable_rerun_hint`
(promoted for the in-crate test mod). `turn.rs` remains the only
in-crate consumer.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
